Add button to profile screen that redirects admin user to admin interface
