+++ frontend/src/components/ProfileComponent.js
@@ -25,7 +25,10 @@ import {
   selectDisplayname,
   selectEmail,
   selectPreferredLanguages,
-  setPreferredLanguages
+  setPreferredLanguages,
+  selectIdToken,
+  selectIsAuthorized,
+  setIsAuthorized
 } from '../redux/UserSlice'
 
 function Profile () {
@@ -34,6 +37,30 @@ function Profile () {
   const email = useSelector(selectEmail)
   const displayName = useSelector(selectDisplayname)
   const preferredLanguages = useSelector(selectPreferredLanguages)
+
+  const userId = useSelector(selectUserid)
+  const idToken = useSelector(selectIdToken)
+  const isAuthorized = useSelector(selectIsAuthorized)
+
+  useEffect(() => {
+    const config = {
+      headers: { Authorization: `Bearer ${idToken}` }
+    }
+
+    axios
+      .get(`http://localhost:3001/user/authorizeAdmin/${userId}`, config)
+      .then((response) => {
+        if (response.status === 200) {
+          dispatch(setIsAuthorized(true))
+        } else {
+          dispatch(setIsAuthorized(false))
+        }
+      })
+      .catch((error) => {
+        console.error('Error checking authorization:', error)
+        dispatch(setIsAuthorized(false))
+      })
+  }, [userId, idToken])
 
   const ALL_LANGUAGES = ['Python', 'Java', 'C']
 
@@ -46,7 +73,6 @@ function Profile () {
     axios
       .get(`http://localhost:3001/user/getLanguage/${userid}`)
       .then((response) => {
-        console.log(response)
         const userLanguages = response.data.languages
         dispatch(setPreferredLanguages(userLanguages))
       })
@@ -145,6 +171,10 @@ function Profile () {
                 change preferred languages
               </Typography>
             </Link>
+
+            {isAuthorized
+              ? (<a href='/admin/questions'><Button variant='outlined'>View questions </Button></a>)
+              : <></>}
 
             <Dialog
               open={isLanguageChangeDialogOpen}
